Fix FIXT 1.1 support
Although FIXT 1.1 is working for outbound messages, it is not working
for inbound messages. The message parser uses a field capacity of eight
bytes for BeginString(8), just like the session did.

Increase the field capacity for BeginString(8) in the message parser.
Define constants for header and trailer field capacity values that are
shared by both the message parser and the session.