Generate alias for CxlRejResponseTo(434) value
The symbolic names generated for enumerations are based on the symbolic
names contained in the FIX Repository. The CxlRejResponseTo(434) values
look as follows:

  Value  Symbolic Name       Description
  -----  ------------------  ----------------------------
      1  OrderCancelRequest  Order Cancel Request
      2  OrderCancel         Order Cancel/Replace Request

The symbolic name 'OrderCancel' for the value that indicates an Order
Cancel/Replace Request message is confusing. Fix this by generating an
alias with the symbolic name 'OrderCancelReplaceRequest'.