philadelphia-core: Remove 'FIXField' class
The 'FIXMessage' class contains an array of 'FIXField' objects, and
the 'FIXField' class contains a tag and a 'FIXValue' object. Remove
the 'FIXField' class and make the 'FIXMessage' class contain two
arrays: one for tags and another for 'FIXValue' objects.

Not only does this improve spatial locality within a 'FIXMessage'
object, it also makes working with the class simpler.