philadelphia-core: Simplify keep-alive mechanism
After sending a TestRequest(1) message, consider the procedure
successful on receiving any data from the other party, not just
a Heartbeat(0) message.

The purpose of this change is to separate the keep-alive mechanism
from the handling of administrative messages, making it easier to
provide a way for applications to implement their own handling of
administrative messages.