philadelphia-core: Improve MsgType(35) handling
When checking for SequenceReset(4), check the first byte of the
MsgType(35) value before checking its length. This avoids one
conditional check for most messages.

Note that the length of the MsgType(35) value could in principle
be zero. In that case, "the first byte of the value" is actually
the SOH character.