philadelphia-core: Remove transitive Joda-Time dependency
Currently, Philadelphia Core depends on Joda-Time 2.10.5. This means
that any application that has a dependency on Philadelphia Core also
has a transitive dependency on Joda-Time 2.10.5.

Replace the transitive Joda-Time 2.10.5 dependency with a Joda-Time
2.x dependency. This makes it possible for applications to manage
their Joda-Time 2.x and Philadelphia Core dependencies independently.

This also means that applications that use Philadelphia Core but do
not yet use Joda-Time 2.x need to add a dependency on Joda-Time 2.x
as well.