philadelphia-core: Fix 'FIXValue#asDate'
This fixes a performance regression in the method by reverting the
changes done since the 1.1.1 release.

In the commit 1f0a188 the method was changed so that it only sets the
date fields and not the time fields. This made the method 82% slower.

In the commit fee8874 the method was changed so that it sets the date
fields one by one rather than all at once. This made the method faster
again, but it was still 62% slower than in the 1.1.1 release.