philadelphia-core: Improve backing store management in 'FIXValue'
Make it possible to specify a backing array for a value container.

Add two new internal attributes: 'start' and 'end'. These are the
start index (inclusive) and the end index (exclusive) for the value
container's slot in the backing array.

The existing internal attributes 'offset' and 'length' retain the
same semantics as before. In particular, 'offset' still refers to
the position of the contained value within the entire backing array
and is not relative to 'start'.

Because 'offset' and 'length' work the same way as before, the value
accessors, such as 'FIXValue#asBoolean', do not change. However, the
value mutators, such as 'FIXValue#setBoolean' do change: the value
container's slot in the backing array now starts at 'start' instead
of 0 and ends at 'end' instead of 'bytes.length'.

The bytecode changes are as follows:

  Method        Before  After  Change
  ------------  ------  -----  ------
  set               41     41       0
  setBoolean        34     45     +11
  setChar           25     36     +11
  setInt           102     99      -3
  setFloat         177    174      -3
  setString         55     68     +13
  setDate           59     80     +21
  setTimeOnly      119    165     +46
  setTimestamp     172    238     +66
  setCheckSum       28     39     +11
  get               65     67      +2