philadelphia-core: Improve buffer capacity management
Add 'FIXConfig#getBufferCapacity()' and deprecate the following methods:

  - 'FIXConfig#getRxBufferCapacity()'
  - 'FIXConfig#getTxBufferCapacity()'

The purpose of this change is to simplify configuration. As the maximum
number of fields and the field capacity already apply to both incoming
and outgoing messages, it makes sense to have one buffer capacity that
applies to both as well.

The deprecated methods will continue working as before until they get
removed in Philadelphia 2.0.0.