philadelphia-core: Add 'FIXTimestamp'
Replace 'MutableDateTime' usage with a 'FIXTimestamp' class, and
remove the Joda-Time 2.x dependency. In addition to removing the
dependency, this improves the performance significantly.

  Benchmark                      Baseline Candidate
  ------------------------------ -------- ---------
  FIXValueBenchmark.asDate           1.00      0.46
  FIXValueBenchmark.asTimeOnly       1.00      0.38
  FIXValueBenchmark.asTimestamp      1.00      0.49
  FIXValueBenchmark.setDate          1.00      0.68
  FIXValueBenchmark.setTimeOnly      1.00      0.44
  FIXValueBenchmark.setTimestamp     1.00      0.31

Note that the benchmarks above include both parsing or formatting
the values to or from bytes as well as converting the timestamps
to or from numbers of milliseconds from the epoch.