philadelphia-core: Clean up test
The test uses blocking I/O to ensure that the implementation under test
receives Heartbeat(0) messages. This is because a Heartbeat(0) message
does not have a callback unlike, for example, a Logon(A) message.

Make the test simpler by replacing blocking I/O with a MsgSeqNum(34)
check. Although a Heartbeat(0) message does not invoke a callback, it
still increments the MsgSeqNum(34) value.