philadelphia-core: Improve 'FIXConnection' interface
Refer to underlying channels rather than underlying socket channels
everywhere.

Document the type bounds for the underlying channel type in the generic
constructor, and retain a reference to 'SocketChannel', still the most
common underlying channel type, in the Javadoc comment.

Prefer the terms "inbound" and "outbound" to the terms "RX" and "TX" in
the API.