philadelphia-core: Add 'FIXKeepAliveStatus'
Signal a heartbeat timeout using the 'FIXConnection#keepAlive()' return
value rather than via the 'FIXConnectionStatusListener' interface.

The purpose of this change is to separate the administrative message
handling implemented in 'FIXConnection#keepAlive()' from that currently
implemented in a private inner class of 'FIXConnection'. All callbacks
in 'FIXConnectionStatusListener' except 'heartbeatTimeout()' are invoked
by the inner class. By making 'FIXConnection#keepAlive()' independent
of 'FIXConnectionStatusListener', 'FIXConnection' in whole will become
independent of it. This will make it possible to move the inner class
out of 'FIXConnection' and allow applications to provide their own
implementations of administrative message handling.